# Faster ChiselPlugin