Use caching to reduce number of expensive lookups. Also make methods private so we can improve them without breaking binary compatibility

I think this could be better, like only doing a single lookup, but I was having trouble getting it working so thought it better to just get this out here. This reduces the runtime of the Chisel plugin in rocket-chip compile by quite a bit (collected with scalacOption `-Ystatistics`):

```
Before:
[info]   typer                       : 1 spans, ()32910.11ms (27.8%)
[info]   chiselcomponent             : 1 spans, ()29308.0ms (24.7%)

After:
[info]   typer                       : 1 spans, ()26130.235ms (42.1%)
[info]   chiselcomponent             : 1 spans, ()6523.072ms (10.5%)
```
Typer included for reference (also shows how noisy the measurements are).
With this change, chiselcomponent is generally 2-4x faster.

### Contributor Checklist

- [ ] Did you add Scaladoc to every public function/method?
- [ ] Did you add at least one test demonstrating the PR?
- [x] Did you delete any extraneous printlns/debugging code?
- [x] Did you specify the type of improvement?
- [NA] Did you add appropriate documentation in `docs/src`?
- [x] Did you state the API impact?
- [x] Did you specify the code generation impact?
- [x] Did you request a desired merge strategy?
- [x] Did you add text to be included in the Release Notes for this change?

#### Type of Improvement

  - performance improvement 

#### API Impact

Some methods that shouldn't have been public in the compiler plugin are now private

#### Backend Code Generation Impact

No impact

#### Desired Merge Strategy

- Squash

#### Release Notes

Improve performance of the Chisel Scalac compiler plugin over that in 3.4.0-RC1

### Reviewer Checklist (only modified by reviewer)
- [ ] Did you add the appropriate labels?
- [ ] Did you mark the proper milestone (3.2.x, 3.3.x, 3.4.0, 3.5.0) ?
- [ ] Did you review?
- [ ] Did you check whether all relevant Contributor checkboxes have been checked?
- [ ] Did you mark as `Please Merge`?